Don't show find-and-replace "Copy Path" item in tree-view context menu
9e53ae0174bda322ee43ed494204ca342a78d782 reduced the specificity of the
selector for the "Copy Path" menu item [A], causing the menu item to
appear in the tree-view's context menu. With the changes in this commit,
the "Copy Path" menu item appears in the find-and-replace context menu
for files and lines, but it does not appear in the tree-view context
menu.

[A]
https://github.com/atom/find-and-replace/commit/9e53ae0174bda322ee43ed494204ca342a78d782#diff-9d7d0803c9fece359b6cd239f42a2e23L31