Simplify some syntax that was more complex than necessary (#149)
The end tag for the block was (?<=>)|(?!;), which could be simplified
to (?!;), which always matches the next character after the begin
block. This makes the size of the block 0, so nothing is actually happening
inside the block, just in the beginCaptures, so I just converted the
beginCaptures to their own matching group.