add typename to component state
It feels a bit dirty to be checking in multiple places, and also that 
both the issue and the pullRequest have a typename. 

I couldn't figure out how to pull the `__typename` up a level in 
graphql, but at least having a single source of truth in state for the 
type of the type of item we're dealing with seems marginally cleaner.