reverse order of comment threads in pr builder
the test was asserting on 3 individual comments, that they rendered in 
the correct positions.

when Katrina implemented grouping comments by thread, she did something 
like: `[...this.props.commentThreads].reverse().map(({rootCommentId, 
comments}) => {`

tbh I don’t really understand why we need to call `reverse` there
but we also need to be calling reverse in the comment builder, so that 
the tests are doing the same thing as the code.

and then there was also a minor problem where we were passing in two 
comments with the same ID, which is now fixed.