use default git cleanup configuration if commit template exists.
in https://github.com/atom/github/issues/1817, there is a bug where we are not stripping commented lines if a git message template exists.

The `verbatim` flag is passed if the commit is made from a mini editor.  The mini editor is meant to be analagous of `git commit -m`, a quick and dirty sort of affair.

As much as I don't want to add special cases here, it seemed like the cleanest to ignore `verbatim` if you have a template set.