Add support for `final` in `catch` parameters (#188)
### Description of the Change

The regex for the `catch` parameters currently does not support `final`, meaning that `final` is highlighted as a type. This change adds an optional group to account for the possibility of the Java `final` modifier.

### Alternate Designs

None, as prepending the existing regex is a simple change.

### Benefits

Correct syntax highlighting when `final` is within `catch` parameters

### Possible Drawbacks

None.

### Applicable Issues

None.