Don't return an unresolved Promise on beforeEach()
The `editor.update()` method returns a Promise, which under some
circumstances (on specs mostly) never gets resolved.

Because of CoffeeScript's implicit returns, this Promise was returned on
the `beforeEach()` method of these tests.

Before https://github.com/atom/atom/pull/18917, any Promise that was
returned by a spec method was just discarded, so this behaviour didn't
cause any issue, but once we merge that PR this test is going to start
failing because jasmine will wait indefinitely until the Promise is
resolved.