Provide a non-render callback prop to Accumulator
React doesn't particularly like it when you call `setState()` or
`forceUpdate()` (or other side-effect-causing functionality) during
`render()`. According to the React docs, `render()` is supposed to be a
pure function. To allow us to trigger side effects as a result of
accumulated result batches arriving, I'm adding a `handleResults` prop
to Accumulator that's triggered outside of the render cycle (during
`componentDidMount` or `loadMore`) that we can use to do these kinds of
actions.