Turn PRCheckoutController into a Relay fragment that uses a render prop
Wrapping PRCheckoutController with a Relay fragment container lets us
declare its data dependencies closest to the place where they're used
and reduce duplication in the containers that nest them. I did have to
change it to use a render prop instead of a `childItem` prop to render
the component that receives the checkoutOp because Relay couldn't track
which fragment belonged to which component beneath the
PRCheckoutController otherwise.