Specialized builder for AggregatedReviewContainer output
The AggregatedReviewsContainer is used in a few places in our React
tree. The data it constructs to pass to its render prop is awkward to
make with the GraphQL-synchronized builders, because it collates the
results from three separate, individually paginated fragments. This is a
specialized builder to make its results easier to mock.