addGutterDecoration -> createGutter
So the original reason the gutter decorations weren't working was because if you try to decorate a gutter that doesn't exist, it silently fails.  Cool.  So create the gutter if it doesn't exist, and this simplifies the logic around adding decorations. we call `editor[this.props.decorateMethod](decorable, opts)` instead of calling gutter.decorateMarker.