Allow for more than one key/value pair per line
The rewritten grammar is courtesy of @50Wilu, who had this to say:

"The problem was that the current language structure couldn't handle
more than one key/value pair per line, which inline tables allow. To fix
that, this diff encapsulates each key pattern so that it'll also match
the corresponding value before finishing the match, and ensuring that
values won't match unless paired with a key. I've added explicit support
for arrays and inline tables so that highlighting for those two don't
regress."

These changes originally broke many of the tests, which I have since
refactored to use `key = value` syntax instead.