Add throttling to the filter metric events
This will ensure that we don't send every single filter event happening
on each keystroke from the user on the fuzzy finder, which would be too
much data.

Adding some throttling could cause some bias (as opposed to do standard
sampling), since we're going to send the first and last filter events
from a search than the intermediate ones, but that should not be an
issue since all these searches should perform equally (and the last one
is actually the most important one for the user).