Load language-test package before activating snippets in all tests
When the snippets package activates, it reads snippets from all of the 
loaded packages in an async code block after loading bundled snippets. 

Previously, tests were loading the `language-test` package after 
activating snippets, which created a race between the async code block 
that read the loaded packages and the loading of this package. Most of 
the time, the jasmine waitsFor block would load the package before the 
async code in the snippets package scanned the loaded package. 
Sometimes, however, it would not. Hopefully this addresses this 
unreliability by always activating the language package before the 
snippets package to avoid the race.