Support for Java 14's Record (#226)
### Description of the Change
Support "record" for Java 14, see #223 . Below are covered cases. 
```
1. basic declaration
record Point() {}

2. having record constructor, meanwhile not breaking member functions
public record Point(int x, int y) {
  public Point {
    // validation
  }
  private void foo() { }
}

3. implemeting interfaces
public record Point(int x) implements IA, IB {}

4. defined inner another class
class A {
  record Point() {}
}
```
<!--

We must be able to understand the design of your change from this description. If we can't get a good idea of what the code will be doing from the description here, the pull request may be closed at the maintainers' discretion. Keep in mind that the maintainer reviewing this PR may not be familiar with or have worked with the code here recently, so please walk us through the concepts.

-->

### Alternate Designs

N/A
### Benefits
Support to hightlight 'record'
### Possible Drawbacks
new rules are limited to 'record' declarations, and are supposed not to causing any no regression.
<!-- What are the possible side-effects or negative impacts of the code change? -->

### Applicable Issues
Closes #223 
<!-- Enter any applicable Issues here -->