Add ❯ and ➜ as prompt characters
The ❯ character is the most common default I've seen for a non-ASCII
prompt character ([1][], [2][], [3][], [4][]) that isn't a Powerline
prompt, so allow it as a prompt character.

Also adding ➜ as the other common one (default in oh-my-zsh etc).

I went through the top-starred GitHub themes [5][], and these were the only
two common ones.

[1]: https://github.com/sindresorhus/pure
[2]: https://github.com/JanDeDobbeleer/oh-my-posh
[3]: https://github.com/ohmyzsh/ohmyzsh/wiki/External-themes
[4]: https://github.com/romkatv/powerlevel10k
[5]: https://github.com/search?o=desc&q=shell+prompt&s=stars&type=Repositories