🎨 Remove the CSON grammar file for GitHub Flavored Markdown
We are removing this file because we will be replacing it with a JSON
file that is refactored into groups. This will allow for better reading
of the the file and additions if needed.

This refactor is not purely cosmetic, the refactor addresses bug that
was raised by a GitHub customer about syntax highlighting not working
for a specific language