Update CI build
### Requirements

* Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
* All new code requires tests to ensure against regressions

### Description of the Change

This PR updates CI build from Travis and Appveyor to GitHub Actions. This is motivated by the following reasons:
- Travis CI API has been having some issues with status propagation, the same applies to Appveyor. 
- Builds on Travis are stuck sometimes.
- Definition is extended to run against ubuntu, macos, and windows.
- Consolidates execution into a single file, there is no disparity between configuration files.

### Alternate Designs

There was another PR that adds GitHub Actions to this repository; however, it was incomplete, e.g. missing lint step, and did not clean up the previous build files.

### Benefits

Fast and reliable builds.

### Possible Drawbacks

Potential updates to Atom builds will have to be handled manually instead of relying on community updates to the installation scripts.

### Applicable Issues

<!-- Enter any applicable Issues here -->