Adding '$' character for highlight-selected
Per this thread at Atom: https://discuss.atom.io/t/double-click-highlights-whole-of-variable-in-php/18625

Double-cliccking a variable is not recognized by highlight-selcted by the original default setting for Non-word characters. Adding a '$' character to the beginning fixes this problem so highlight-selected pics it up.

If there is another way to solve this problem so double-clicked variables highlight, by all means, do that instead.